+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.22.3
 ---------------------------
 * Fix #1220 Use relative path for public/cache/ for JS and CSS files
 * Implement XEP-0425: Message Moderation
+* Improve the avatars picture quality upload, 512x512
 
 v0.22.2
 ---------------------------